requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
anthropic
streamlit
pandas
requests